Combine C++ sources to reduce package size
